requirements.txt
certifi==2021.5.30
click==7.1.2
cycler==0.11.0
kiwisolver==1.3.1
matplotlib==3.3.4
mysql-connector-python==8.0.27
numpy==1.19.5
pandas==1.1.5
Pillow==8.4.0
protobuf==3.19.1
pyparsing==3.0.6
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
sip==4.19.8
six==1.16.0
TA-Lib @ file:///C:/Users/lsw_owen/Desktop/211221_Inflearn_Trading/TA_Lib-0.4.19-cp36-cp36m-win32.whl
wincertstore==0.2
